unify some apple code
