+++ src/Maui/DrawnUi/Platforms/Apple/Handlers/SKGLViewHandlerRetained.Apple.cs
@@ -87,16 +87,7 @@ namespace DrawnUi.Views
             if (handler?.PlatformView == null)
                 return;
 
-            // EXPERIMENT: view-driven pacing. On-demand mode (Paused + SetNeedsDisplay)
-            // has uneven cadence during animations: the next frame is scheduled by an
-            // invalidate issued from INSIDE the current draw, and CoreAnimation honors
-            // it either this vsync or the next — measured 33/66ms alternation at MaxFps=30
-            // while the tick loop itself was a perfect 33.4ms. Continuous MTKView at the
-            // capped rate makes the view the single pacer: a draw EVERY divided vsync,
-            // animators stepped inside a guaranteed-cadence draw pass.
-            handler.PlatformView.Paused = false;
-            handler.PlatformView.EnableSetNeedsDisplay = false;
-            handler.PlatformView.PreferredFramesPerSecond = Super.MaxFps > 0 ? Super.MaxFps : 0;
+            Super.ApplyMetalViewPacing(handler.PlatformView, Super.MaxFps);
         }
 
         public static void MapEnableTouchEvents(SKGLViewHandlerRetained handler, ISKGLView view)
