+++ src/Maui/DrawnUi/Platforms/MacCatalyst/Super.Mac.cs
@@ -109,14 +109,6 @@ namespace DrawnUi.Draw
 
         #endregion
 
-        public static float GetDisplayRefreshRate(float fallback)
-        {
-            return (int)UIScreen.MainScreen.MaximumFramesPerSecond;
-        }
-
-        public static float RefreshRate { get; protected set; }
-
-
         public static void Init()
         {
             if (Initialized)
@@ -211,7 +203,7 @@ namespace DrawnUi.Draw
             // Display link callback reads MaxFps dynamically, no action needed there.
             // Update looper fps if it's being used instead of CADisplayLink.
             Looper?.SetTargetFps(fps > 0 ? fps : RefreshRate);
-                    UpdateRegisteredMetalViewsPreferredFramesPerSecond(fps);
+            UpdateRegisteredMetalViewsPacing(fps);
         }
 
         public static event EventHandler OnFrame;
@@ -219,95 +211,6 @@ namespace DrawnUi.Draw
         static CADisplayLink _displayLink;
         static double _lastDisplayTimestamp;
 
-        /// <summary>
-        /// Opens web link in native browser
-        /// </summary>
-        /// <param name="link"></param>
-
-        public static void OpenLink(string link)
-        {
-            try
-            {
-                var url = new NSUrl(link);
-                var res = UIApplication.SharedApplication.OpenUrl(url);
-            }
-            catch (Exception e)
-            {
-                Super.Log(e);
-            }
-        }
-
-        /// <summary>
-        /// Lists assets inside the Resources/Raw subfolder
-        /// </summary>
-        /// <param name="subfolder"></param>
-        /// <returns></returns>
-        public static IEnumerable<string> ListResources(string subfolder)
-        {
-            NSBundle mainBundle = NSBundle.MainBundle;
-            string resourcesPath = mainBundle.ResourcePath;
-            string subfolderPath = Path.Combine(resourcesPath, subfolder);
-
-            if (Directory.Exists(subfolderPath))
-            {
-                string[] files = Directory.GetFiles(subfolderPath);
-                return files.Select(Path.GetFileName).ToList();
-            }
-            else
-            {
-                return new List<string>();
-            }
-        }
-
-        public static async Task<byte[]> CaptureScreenshotAsync()
-        {
-            UIWindow window;
-            if (UIDevice.CurrentDevice.CheckSystemVersion(13, 0))
-            {
-                window = UIApplication.SharedApplication.Windows.First();
-                if (window.GetType().Name.Contains("Popup"))
-                {
-                    var maybe = UIApplication.SharedApplication.Windows.FirstOrDefault(x => x != window);
-                    if (maybe != null)
-                    {
-                        window = maybe;
-                    }
-                }
-            }
-            else
-            {
-                window = UIApplication.SharedApplication.KeyWindow;
-            }
-
-            UIGraphics.BeginImageContextWithOptions(window.Bounds.Size, false, UIScreen.MainScreen.Scale);
-            window.Layer.RenderInContext(UIGraphics.GetCurrentContext());
-            UIImage image = UIGraphics.GetImageFromCurrentImageContext();
-            UIGraphics.EndImageContext();
-            using (NSData data = image.AsPNG())
-            {
-                var bytes = new byte[data.Length];
-                Marshal.Copy(data.Bytes, bytes, 0, Convert.ToInt32(data.Length));
-                return bytes;
-            }
-
-        }
-
-        private static bool _keepScreenOn;
-        /// <summary>
-        /// Prevents display from auto-turning off  Everytime you set this the setting will be applied.
-        /// </summary>
-        public static bool KeepScreenOn
-        {
-            get
-            {
-                return _keepScreenOn;
-            }
-            set
-            {
-                UIApplication.SharedApplication.IdleTimerDisabled = value;
-                _keepScreenOn = value;
-            }
-        }
 
     }
